requirements.txt
fastapi
uvicorn
pymongo
python-jose
python-dotenv
pydantic[email]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4